refactor(api): clean up public API exports
Remove internal implementation details from public API:
- Remove Dispatcher, StateMap, RenderConfig from app exports
- Remove ComponentId, StateExt from component exports (kept MessageExt for macros)
- Remove Patch, diff, RenderNode, VDom, VNode, wrap_text exports
- Remove render_node_to_buffer from app module exports
- Update internal imports to use direct paths instead of public exports

Keep macro-required exports:
- MessageExt (used by #[update] macro for downcasting)
- Spacing, Overflow, Position (used by node! macro)

All tests pass and examples compile successfully.
