+++ rxtui/lib/app/mod.rs
@@ -8,7 +8,5 @@ pub mod renderer;
 // Exports
 //--------------------------------------------------------------------------------------------------
 
-pub use config::RenderConfig;
-pub use context::{Context, Dispatcher, StateMap};
+pub use context::Context;
 pub use core::App;
-pub use renderer::render_node_to_buffer;
